@@ -8,7 +8,10 @@ exclude = [
   "examples",
   "rules",
   ".env*",
-  "review-context-*.md"
+  "review-context-*.md",
+  "code-review-context-*.md",
+  "code-review-comprehensive-feedback-*.md",
+  "meta-prompt-*.md"
 ]
 
 [tool.hatch.build.targets.wheel]
@@ -18,26 +21,27 @@ exclude = [
   "examples", 
   "rules",
   ".env*",
-  "review-context-*.md"
+  "review-context-*.md",
+  "code-review-context-*.md",
+  "code-review-comprehensive-feedback-*.md",
+  "meta-prompt-*.md"
 ]
 
 [project]
-name = "task-list-code-review-mcp"
-version = "0.2.1"
-description = "MCP server for AI-powered task list code reviews with PRD context"
+name = "gemini-code-review-mcp"
+version = "0.3.9"
+description = "MCP server for AI-powered code reviews using Google Gemini with contextual awareness"
 authors = [
     {name = "Nico Bailon", email = "nico604@pm.me"},
 ]
 readme = "README.md"
 license = {text = "MIT"}
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 classifiers = [
     "Development Status :: 4 - Beta",
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -50,9 +54,9 @@ dependencies = [
 ]
 
 [project.urls]
-Homepage = "https://github.com/nicobailon/task-list-code-review-mcp"
-Repository = "https://github.com/nicobailon/task-list-code-review-mcp"
-Issues = "https://github.com/nicobailon/task-list-code-review-mcp/issues"
+Homepage = "https://github.com/nicobailon/gemini-code-review-mcp"
+Repository = "https://github.com/nicobailon/gemini-code-review-mcp"
+Issues = "https://github.com/nicobailon/gemini-code-review-mcp/issues"
 
 [project.optional-dependencies]
 dev = [
@@ -61,17 +65,30 @@ dev = [
     "pytest-asyncio>=0.21.0",
     "black",
     "isort",
-    "flake8",
+    "pyright",
+]
+yaml = [
+    "PyYAML>=6.0.0",
 ]
 
 [project.scripts]
-task-list-code-review-mcp = "src.server:main"
-generate-code-review = "src.generate_code_review_context:cli_main"
-review-with-ai = "src.ai_code_review:main"
+# MCP Server
+gemini-code-review-mcp = "src.server:main"
+# CLI Tools
+generate-code-review = "src.cli_main:main"
+generate-meta-prompt = "src.meta_prompt_generator:main"
+# Semantic shortcuts (when installed)
+code-review = "src.cli_main:main"
+
+[dependency-groups]
+dev = [
+    "psutil>=7.0.0",
+    "pytest>=8.3.5",
+]
 
 [tool.black]
 line-length = 88
-target-version = ['py38']
+target-version = ['py310', 'py311', 'py312']
 
 [tool.isort]
 profile = "black"
